Remove room router, put consts and funcs to separate dirs
