+++ constants/bonuses.js
@@ -1,0 +1,31 @@
+wordBonuses = {
+  0: { 0: 3, 7: 3, 14: 3 },
+  1: { 1: 2, 13: 2 },
+  2: { 2: 2, 12: 2 },
+  3: { 3: 2, 11: 2 },
+  4: { 4: 2, 10: 2 },
+  7: { 0: 3, 14: 3 },
+  10: { 4: 2, 10: 2 },
+  11: { 3: 2, 11: 2 },
+  12: { 2: 2, 12: 2 },
+  13: { 1: 2, 13: 2 },
+  14: { 0: 3, 7: 3, 14: 3 },
+};
+
+letterBonuses = {
+  0: { 3: 2, 11: 2 },
+  1: { 5: 3, 9: 3 },
+  2: { 6: 2, 8: 2 },
+  3: { 0: 2, 7: 2, 14: 2 },
+  5: { 1: 3, 13: 3 },
+  6: { 2: 2, 6: 2, 8: 2, 12: 2 },
+  7: { 3: 2, 11: 2 },
+  8: { 2: 2, 6: 2, 8: 2, 12: 2 },
+  9: { 1: 3, 13: 3 },
+  11: { 0: 2, 7: 2, 14: 2 },
+  12: { 6: 2, 8: 2 },
+  13: { 5: 3, 9: 3 },
+  14: { 3: 2, 11: 2 },
+};
+
+module.exports = { wordBonuses, letterBonuses };
