+++ routers/game.js
@@ -1,459 +1,203 @@
-const sequelize = require("sequelize");
-
 const { Router } = require("express");
 const authMiddleware = require("../auth/middleware");
 const { room, user, game } = require("../models");
-const lettersSets = require("../letterSets");
+const lettersSets = require("../constants/letterSets");
+const {
+  shuffle,
+  getNextTurn,
+  updateGameLetters,
+  getHorizontalWords,
+  rotate,
+  turnWordsAndScore,
+  substract,
+  giveLetters,
+  getResult,
+} = require("../services/game");
 
-const wordBonuses = {
-  0: { 0: 3, 7: 3, 14: 3 },
-  1: { 1: 2, 13: 2 },
-  2: { 2: 2, 12: 2 },
-  3: { 3: 2, 11: 2 },
-  4: { 4: 2, 10: 2 },
-  7: { 0: 3, 14: 3 },
-  10: { 4: 2, 10: 2 },
-  11: { 3: 2, 11: 2 },
-  12: { 2: 2, 12: 2 },
-  13: { 1: 2, 13: 2 },
-  14: { 0: 3, 7: 3, 14: 3 },
-};
-
-const letterBonuses = {
-  0: { 3: 2, 11: 2 },
-  1: { 5: 3, 9: 3 },
-  2: { 6: 2, 8: 2 },
-  3: { 0: 2, 7: 2, 14: 2 },
-  5: { 1: 3, 13: 3 },
-  6: { 2: 2, 6: 2, 8: 2, 12: 2 },
-  7: { 3: 2, 11: 2 },
-  8: { 2: 2, 6: 2, 8: 2, 12: 2 },
-  9: { 1: 3, 13: 3 },
-  11: { 0: 2, 7: 2, 14: 2 },
-  12: { 6: 2, 8: 2 },
-  13: { 5: 3, 9: 3 },
-  14: { 3: 2, 11: 2 },
-};
-
-function shuffle(arr) {
-  return arr.sort(() => Math.random() - 0.5);
-}
-
-function getNextTurn(game) {
-  return (game.turn + 1) % game.turnOrder.length;
-}
-
-function updateGameLetters(game) {
-  const currentUserId = game.turnOrder[game.turn];
-  const currentUserLetters = game.letters[currentUserId];
-  let requiredQty = 7 - currentUserLetters.length;
-  if (game.letters.pot.length < requiredQty) {
-    requiredQty = game.letters.pot.length;
-  }
-  let newLetters = [];
-  while (newLetters.length !== requiredQty) {
-    newLetters.push(
-      game.letters.pot.splice(
-        Math.floor(Math.random() * game.letters.pot.length),
-        1
-      )[0]
-    );
-  }
-  const updatedUserLetters = currentUserLetters.concat(newLetters);
-  const updatedGameLetters = {
-    ...game.letters,
-    [currentUserId]: updatedUserLetters,
-  };
-  return updatedGameLetters;
-}
-// как найти слово? Находим букву, двигаемся влево от нее по обеим доскам, пока не найдем пустую
-// клетку. Затем двигаемся вправо от нее по обеим доскам пока не найдем пустую клетку. Если
-// между пустыми клетками 1 клетка, то это не слово. В противном случае - слово. Проверяем,
-// на что умножать слово. Добавляем стоимость всех букв, проверяя умножение каждой буквы.
-// умножаем слово. Прибавляем к очкам. Как найти следующее слово и не повториться?
-// слово это вектор от ху до ху.
-// слова проверяем только для тех букв, которые не ходятся внутри этих векторов. Но только по
-// горзизонтали. Слова по вертикали ищем только после того,как нашли все слова по горизонтали.
-function getHorizontalWords(board, previousBoard) {
-  return board.reduce((boardWords, row, yIndex) => {
-    return boardWords.concat(
-      row.reduce((lineWords, cell, xIndex) => {
-        if (
-          cell &&
-          !previousBoard[yIndex][xIndex] &&
-          // this cell is not counted in words yet
-          !lineWords.find(
-            (word) =>
-              word.y == yIndex && word.x <= xIndex && word.x + word.len > xIndex
-          )
-        ) {
-          // start to search for the beginning of the word
-          // move left while xIndex < 0 or cell is empty
-          let leftIndex = xIndex;
-          //indefinite loop
-          for (;;) {
-            leftIndex--;
-            if (
-              leftIndex < 0 ||
-              (!board[yIndex][leftIndex] && !previousBoard[yIndex][leftIndex])
-            ) {
-              leftIndex++;
-              break;
-            }
-          }
-          // move right till xIndex >= 15 or cell is empty
-          let rightIndex = xIndex;
-          //indefinite loop
-          for (;;) {
-            rightIndex++;
-            if (
-              rightIndex >= row.length ||
-              (!board[yIndex][rightIndex] && !previousBoard[yIndex][rightIndex])
-            ) {
-              break;
-            }
-          }
-          const len = rightIndex - leftIndex;
-          if (len >= 2) {
-            lineWords.push({
-              y: yIndex,
-              x: leftIndex,
-              len: len,
-              word: board[yIndex].slice(leftIndex, rightIndex),
-            });
-          }
-        }
-        return lineWords;
-      }, [])
-    );
-  }, []);
-}
-
-function rotate(board) {
-  return Array(15)
-    .fill(null)
-    .map((_, index) => board.map((row) => row[index]));
-}
-
-function countWordScore(wordMultiplier, wordObject, previousBoard, values) {
-  return (
-    wordMultiplier *
-    wordObject.word.reduce((wordScore, letter, index) => {
-      let letterMultiplier = 1;
-      if (
-        letterBonuses[wordObject.y] &&
-        letterBonuses[wordObject.y][wordObject.x + index] &&
-        // check if it is not a letter of previous player
-        !previousBoard[wordObject.y][wordObject.x + index]
-      ) {
-        letterMultiplier = letterBonuses[wordObject.y][wordObject.x + index];
-      }
-      return wordScore + values[letter[0]] * letterMultiplier;
-    }, 0)
-  );
-}
-
-function turnWordsAndScore(board, previousBoard, bonus15, values) {
-  const horizontalWords = getHorizontalWords(board, previousBoard);
-  const rotatedBoard = rotate(board);
-  const rotatedPreviousBoard = rotate(previousBoard);
-  const verticalWords = getHorizontalWords(rotatedBoard, rotatedPreviousBoard);
-  const horizontalTurn = horizontalWords.reduce(
-    (turn, wordObject) => {
-      let wordMultiplier = 0;
-      for (let i = wordObject.x; i < wordObject.x + wordObject.len; i++) {
-        if (
-          wordBonuses[wordObject.y] &&
-          wordBonuses[wordObject.y][i] &&
-          // check if it is not a letter of previous player
-          !previousBoard[wordObject.y][i]
-        ) {
-          wordMultiplier += wordBonuses[wordObject.y][i];
-        }
-      }
-      if (wordMultiplier === 0) {
-        wordMultiplier = 1;
-      }
-      const wordScore = countWordScore(
-        wordMultiplier,
-        wordObject,
-        previousBoard,
-        values
-      );
-      turn.score += wordScore;
-      turn.words.push({
-        [wordObject.word
-          .map((letter) => {
-            if (letter[0] === "*") {
-              return letter[1];
-            } else {
-              return letter;
-            }
-          })
-          .join("")]: wordScore,
-      });
-      return turn;
-    },
-    { words: [], score: 0 }
-  );
-  const verticalTurn = verticalWords.reduce(
-    (turn, wordObject) => {
-      let wordMultiplier = 0;
-      for (let i = wordObject.x; i < wordObject.x + wordObject.len; i++) {
-        if (
-          wordBonuses[wordObject.y] &&
-          wordBonuses[wordObject.y][i] &&
-          // check if it is not a letter of previous player
-          !rotate(previousBoard)[wordObject.y][i]
-        ) {
-          wordMultiplier += wordBonuses[wordObject.y][i];
-        }
-      }
-      if (wordMultiplier === 0) {
-        wordMultiplier = 1;
-      }
-      const wordScore = countWordScore(
-        wordMultiplier,
-        wordObject,
-        rotate(previousBoard),
-        values
-      );
-      turn.score += wordScore;
-      turn.words.push({
-        [wordObject.word
-          .map((letter) => {
-            if (letter[0] === "*") {
-              return letter[1];
-            } else {
-              return letter;
-            }
-          })
-          .join("")]: wordScore,
-      });
-      return turn;
-    },
-    { words: [], score: 0 }
-  );
-  let bonus = 0;
-  if (bonus15) {
-    bonus += 15;
-  }
-  const turn = {
-    words: horizontalTurn.words.concat(verticalTurn.words),
-    score: horizontalTurn.score + verticalTurn.score + bonus,
-  };
-  return turn;
-}
-
-// extract letters from all letters
-function substract(arr, subarr) {
-  const tempSubarr = subarr.slice().sort();
-  const tempArr = arr.slice().sort();
-  return tempArr.reduce(
-    (acc, letter) => {
-      if (acc.i === tempSubarr.length) {
-        acc.letters.push(letter);
-        return acc;
-      }
-      if (letter === tempSubarr[acc.i]) {
-        acc.i++;
-        return acc;
-      }
-      acc.letters.push(letter);
-      return acc;
-    },
-    { i: 0, letters: [] }
-  ).letters;
-}
-
-function giveLetters(bag, userLetters, lettersToChange) {
-  const tempBag = shuffle(bag.slice().concat(lettersToChange));
-  const requiredQty = lettersToChange.length;
-  let newLetters = [];
-  while (newLetters.length !== requiredQty) {
-    newLetters.push(
-      tempBag.splice(Math.floor(Math.random() * tempBag.length), 1)[0]
-    );
-  }
-  const updatedUserLetters = userLetters.concat(newLetters);
-  return { bag: tempBag, userLetters: updatedUserLetters };
-}
-
-function getResult(score, turns, userIds) {
-  const winScore = Object.keys(score).reduce(
-    (acc, user) => {
-      if (score[user] > 0 && score[user] > acc[0].score) {
-        acc = [{ score: score[user], user: user }];
-      } else if (score[user] === acc[0].score) {
-        acc.push({ score: score[user], user: user });
-      }
-      return acc;
-    },
-    [{ score: 0 }]
-  );
-  let winner = [];
-  if (winScore[0].score > 0) {
-    winner = winScore.map((el) => el.user);
-  }
-  const longestWord = turns.reduce((acc, turn) => {
-    turn.words.forEach((word) => {
-      if (
-        acc.length === 0 ||
-        Object.keys(word)[0].length > acc[0].word.length
-      ) {
-        acc = [{ word: Object.keys(word)[0], user: turn.user }];
-      } else if (Object.keys(word)[0].length === acc[0].word.length) {
-        acc.push({ word: [Object.keys(word)[0]], user: turn.user });
-      }
-    });
-    return acc;
-  }, []);
-  const maxScoreWord = turns.reduce((acc, turn) => {
-    turn.words.forEach((word) => {
-      if (acc.length === 0 || Object.values(word)[0] > acc[0].value) {
-        acc = [
-          {
-            word: Object.keys(word)[0],
-            value: Object.values(word)[0],
-            user: turn.user,
-          },
-        ];
-      } else if (Object.values(word)[0] === acc[0].value) {
-        acc.push({
-          word: [Object.keys(word)[0]],
-          value: Object.values(word)[0],
-          user: turn.user,
-        });
-      }
-    });
-    return acc;
-  }, []);
-  const bestTurnByCount = turns.reduce((acc, turn) => {
-    if (acc.length === 0 || turn.words.length > acc[0].qty) {
-      acc = [{ qty: turn.words.length, turn, user: turn.user }];
-    } else if (turn.words.length === acc[0].qty) {
-      acc.push({ qty: turn.words.length, turn, user: turn.user });
-    }
-    return acc;
-  }, []);
-  const bestTurnByValue = turns.reduce((acc, turn) => {
-    if (acc.length === 0 || turn.score > acc[0].score) {
-      acc = [{ score: turn.score, turn, user: turn.user }];
-    } else if (turn.score === acc[0].score) {
-      acc.push({ score: turn.score, turn, user: turn.user });
-    }
-    return acc;
-  }, []);
-
-  const neverChangedLetters = userIds.filter(
-    (user) => !turns.some((turn) => turn.changedLetters && turn.user === user)
-  );
-  return {
-    winner,
-    longestWord,
-    maxScoreWord,
-    bestTurnByCount,
-    bestTurnByValue,
-    neverChangedLetters,
-  };
-}
-
-function factory(gameStream, roomStream) {
+function factory(gameStream, lobbyStream) {
   const router = new Router();
 
-  router.post("/start", authMiddleware, async (req, res, nxt) => {
-    const roomId = req.body.roomId;
+  router.post("/create", authMiddleware, async (req, res, next) => {
+    const currentUser = req.user;
+    const { maxPlayers, language } = req.body;
     try {
-      const currentRoom = await room.findByPk(roomId, {
-        include: [
-          {
-            model: user,
-            as: "users",
-            attributes: ["id", "name"],
+      const currentGame = await game.create({ maxPlayers, language });
+      const newGameId = currentGame.id;
+      let players = [];
+      let phase = "waiting";
+      if (req.body.players) {
+        players = await user.findAll({
+          where: {
+            id: req.body.players,
           },
-          {
-            model: game,
-            required: false,
-            attributes: ["id", "phase", "turnOrder", "turn", "validated"],
-            where: {
-              phase: {
-                [sequelize.Op.not]: "finished",
-              },
-            },
-          },
-        ],
-      });
-      // create new game only if there is no unfinished game in this room
-      if (currentRoom.games.length === 0) {
-        const turnOrder = shuffle(currentRoom.users.map((user) => user.id));
-        const set = lettersSets[currentRoom.language].letters;
-        // give letters to players
-        const lettersForGame = shuffle(set);
-        let acc = { pot: lettersForGame.slice() };
-        const letters = currentRoom.users.reduce((acc, user) => {
-          if (!acc[user.id]) {
-            acc[user.id] = [];
-          }
-          while (acc[user.id].length !== 7) {
-            acc[user.id].push(
-              acc.pot.splice(Math.floor(Math.random() * acc.pot.length), 1)[0]
-            );
-          }
-          return acc;
-        }, acc);
-        const score = currentRoom.users.reduce((acc, user) => {
-          acc[user.id] = 0;
-          return acc;
-        }, {});
-        const currentGame = await game.create({
-          turnOrder,
-          letters,
-          roomId,
-          score,
-          turns: [],
-          result: {},
-          language: currentRoom.language,
         });
-        await currentGame.setUsers(currentRoom.users);
-        await currentRoom.update({ phase: "started" });
-        const action = {
-          type: "GAME_UPDATED",
-          payload: { gameId: currentGame.id, currentGame },
-        };
-        const string = JSON.stringify(action);
-        gameStream.send(string);
-        // this response is important
-        res.send(currentGame);
+        phase = "ready";
+      } else {
+        players = [currentUser];
       }
-
-      const updatedRoom = await room.findByPk(roomId, {
+      await currentGame.setUsers(players);
+      await currentGame.update({ phase });
+      const updatedGame = await game.findByPk(newGameId, {
+        attributes: [
+          "id",
+          "phase",
+          "turnOrder",
+          "turn",
+          "validated",
+          "language",
+          "maxPlayers",
+        ],
         include: [
           {
             model: user,
             as: "users",
             attributes: ["id", "name"],
           },
+        ],
+      });
+      const action = {
+        type: "NEW_GAME",
+        payload: updatedGame,
+      };
+      const string = JSON.stringify(action);
+
+      lobbyStream.send(string);
+      res.send(updatedGame);
+    } catch (error) {
+      next(error);
+    }
+  });
+
+  router.put("/join", authMiddleware, async (req, res, next) => {
+    const currentUser = req.user;
+    const gameId = req.body.gameId;
+    try {
+      const currentGame = await game.findByPk(gameId, {
+        include: [
           {
-            model: game,
-            required: false,
-            attributes: ["id", "phase", "turnOrder", "turn", "validated"],
-            where: {
-              phase: {
-                [sequelize.Op.not]: "finished",
-              },
-            },
+            model: user,
+            as: "users",
+            attributes: ["id", "name"],
+          },
+        ],
+      });
+      const updatedUsers = currentGame.users.concat([currentUser]);
+      if (updatedUsers.length === currentGame.maxPlayers) {
+        await currentGame.update({
+          phase: "ready",
+        });
+      }
+      await currentGame.setUsers(updatedUsers);
+      const updatedGame = await game.findByPk(gameId, {
+        attributes: [
+          "id",
+          "phase",
+          "turnOrder",
+          "turn",
+          "validated",
+          "language",
+          "maxPlayers",
+        ],
+        include: [
+          {
+            model: user,
+            as: "users",
+            attributes: ["id", "name"],
+          },
+        ],
+      });
+      const actionForLobby = {
+        type: "UPDATED_GAME_IN_LOBBY",
+        payload: updatedGame,
+      };
+      const actionForLobbyString = JSON.stringify(actionForLobby);
+      lobbyStream.send(actionForLobbyString);
+      const actionForGame = {
+        type: "GAME_UPDATED",
+        payload: { gameId, game: updatedGame },
+      };
+      const actionForGameString = JSON.stringify(actionForGame);
+      gameStream.send(actionForGameString);
+      res.send(actionForGameString);
+    } catch (error) {
+      next(error);
+    }
+  });
+
+  router.post("/start", authMiddleware, async (req, res, nxt) => {
+    const gameId = req.body.gameId;
+    try {
+      const currentGame = await game.findByPk(gameId, {
+        include: [
+          {
+            model: user,
+            as: "users",
+            attributes: ["id", "name"],
+          },
+        ],
+      });
+      const turnOrder = shuffle(currentGame.users.map((user) => user.id));
+      const set = lettersSets[currentGame.language].letters;
+      // give letters to players
+      const lettersForGame = shuffle(set);
+      let acc = { pot: lettersForGame.slice() };
+      const letters = currentGame.users.reduce((acc, user) => {
+        if (!acc[user.id]) {
+          acc[user.id] = [];
+        }
+        while (acc[user.id].length !== 7) {
+          acc[user.id].push(
+            acc.pot.splice(Math.floor(Math.random() * acc.pot.length), 1)[0]
+          );
+        }
+        return acc;
+      }, acc);
+      const score = currentGame.users.reduce((acc, user) => {
+        acc[user.id] = 0;
+        return acc;
+      }, {});
+      await currentGame.update({
+        turnOrder,
+        letters,
+        score,
+        turns: [],
+        result: {},
+        phase: "turn",
+      });
+      const action = {
+        type: "GAME_UPDATED",
+        payload: { gameId, game: currentGame },
+      };
+      const string = JSON.stringify(action);
+      gameStream.send(string);
+      // this response is important
+      res.send(currentGame);
+
+      const updatedGameForLobby = await game.findByPk(gameId, {
+        attributes: [
+          "id",
+          "phase",
+          "turnOrder",
+          "turn",
+          "validated",
+          "language",
+          "maxPlayers",
+        ],
+        include: [
+          {
+            model: user,
+            as: "users",
+            attributes: ["id", "name"],
           },
         ],
       });
 
-      updatedRoom.dataValues.game = updatedRoom.dataValues.games[0];
-      delete updatedRoom.dataValues.games;
       const action2 = {
-        type: "UPDATED_ROOM",
-        payload: updatedRoom,
+        type: "UPDATED_GAME_IN_LOBBY",
+        payload: updatedGameForLobby,
       };
       const string2 = JSON.stringify(action2);
-      roomStream.send(string2);
+      lobbyStream.send(string2);
     } catch (error) {
       nxt(error);
     }
@@ -473,7 +217,7 @@ function factory(gameStream, roomStream) {
       });
       const action = {
         type: "GAME_UPDATED",
-        payload: { gameId: gameId, game: currentGame },
+        payload: { gameId, game: currentGame },
       };
       const string = JSON.stringify(action);
       gameStream.init(req, res);
@@ -529,25 +273,12 @@ function factory(gameStream, roomStream) {
                 },
               ],
             });
-            await currentGame.room.update({
-              phase: "ready",
-            });
-            const updatedRoom = await room.findByPk(currentGame.roomId, {
-              include: [
-                {
-                  model: user,
-                  as: "users",
-                  attributes: ["id", "name"],
-                },
-                game,
-              ],
-            });
             const action2 = {
-              type: "UPDATED_ROOM",
-              payload: updatedRoom,
+              type: "DELETE_GAME_IN_LOBBY",
+              payload: gameId,
             };
             const string2 = JSON.stringify(action2);
-            roomStream.send(string2);
+            lobbyStream.send(string2);
           } else {
             await currentGame.update({
               previousBoard: currentGame.board,
@@ -603,9 +334,9 @@ function factory(gameStream, roomStream) {
           if (ok) {
             // check if all letters in turn were in user's hand
             // TODO: take possible duplicates into account
-            // TODO: rewrite, remove returns
+            // TODO: rewrite, it doesn't work because of wild cards from board
             // if (
-            //   userBoard.some((row) =>
+            //   !userBoard.some((row) =>
             //     row.some((letter) => {
             //       console.log(letter);
             //       return (
@@ -621,49 +352,33 @@ function factory(gameStream, roomStream) {
             //     })
             //   )
             // ) {
-            //   return next(`Invalid letters used`);
-            // }
+
             // check if all letters are placed on empty cells
             if (
-              userBoard.some((row, y) =>
+              !userBoard.some((row, y) =>
                 row.some(
                   (letter, x) =>
                     letter !== null && currentGame.board[y][x] !== null
                 )
               )
             ) {
-              return next("Cell is already occupied");
-            }
-            // add user letters to current board
-            const newBoard = currentGameBoard.map((row, y) =>
-              row.map((cell, x) => {
-                if (cell === null && userBoard[y][x] !== null) {
-                  return userBoard[y][x];
-                } else {
-                  return cell;
-                }
-              })
-            );
+              // add user letters to current board
+              const newBoard = currentGameBoard.map((row, y) =>
+                row.map((cell, x) => {
+                  if (cell === null && userBoard[y][x] !== null) {
+                    return userBoard[y][x];
+                  } else {
+                    return cell;
+                  }
+                })
+              );
 
-            // check if there is any duplicated word
-            //if no, proceed with updating game and sending it in stream
-            // if yes, don't update the game, send action with error as a response only
-            const hWords = getHorizontalWords(newBoard, currentGame.board).map(
-              (wordObject) =>
-                wordObject.word
-                  .map((letter) => {
-                    if (letter[0] === "*") {
-                      return letter[1];
-                    } else {
-                      return letter;
-                    }
-                  })
-                  .join("")
-            );
-            const words = hWords.concat(
-              getHorizontalWords(
-                rotate(newBoard),
-                rotate(currentGame.board)
+              // check if there is any duplicated word
+              //if no, proceed with updating game and sending it in stream
+              // if yes, don't update the game, send action with error as a response only
+              const hWords = getHorizontalWords(
+                newBoard,
+                currentGame.board
               ).map((wordObject) =>
                 wordObject.word
                   .map((letter) => {
@@ -674,61 +389,78 @@ function factory(gameStream, roomStream) {
                     }
                   })
                   .join("")
-              )
-            );
+              );
+              const words = hWords.concat(
+                getHorizontalWords(
+                  rotate(newBoard),
+                  rotate(currentGame.board)
+                ).map((wordObject) =>
+                  wordObject.word
+                    .map((letter) => {
+                      if (letter[0] === "*") {
+                        return letter[1];
+                      } else {
+                        return letter;
+                      }
+                    })
+                    .join("")
+                )
+              );
 
-            const duplicatedWords = words.filter((word) => {
-              return currentGame.turns.some((turn) => {
-                if (turn.words.length > 0) {
-                  const listOfWords = turn.words.map(
-                    (wordObject) => Object.keys(wordObject)[0]
+              const duplicatedWords = words.filter((word) => {
+                return currentGame.turns.some((turn) => {
+                  if (turn.words.length > 0) {
+                    const listOfWords = turn.words.map(
+                      (wordObject) => Object.keys(wordObject)[0]
+                    );
+                    return listOfWords.includes(word);
+                  }
+                });
+              });
+              if (duplicatedWords && duplicatedWords.length > 0) {
+                const action = {
+                  type: "DUPLICATED_WORDS",
+                  payload: duplicatedWords,
+                };
+                res.send(JSON.stringify(action));
+                return;
+              } else {
+                // remove those letters from user hand
+                const putLetters = userBoard.reduce((acc, row) => {
+                  return acc.concat(
+                    row.reduce((accum, cell) => {
+                      if (cell) {
+                        accum.push(cell[0]);
+                      }
+                      return accum;
+                    }, [])
                   );
-                  return listOfWords.includes(word);
-                }
-              });
-            });
-            if (duplicatedWords && duplicatedWords.length > 0) {
-              const action = {
-                type: "DUPLICATED_WORDS",
-                payload: duplicatedWords,
-              };
-              res.send(JSON.stringify(action));
-              return;
-            } else {
-              // remove those letters from user hand
-              const putLetters = userBoard.reduce((acc, row) => {
-                return acc.concat(
-                  row.reduce((accum, cell) => {
-                    if (cell) {
-                      accum.push(cell[0]);
-                    }
-                    return accum;
-                  }, [])
-                );
-              }, []);
+                }, []);
 
-              // extract put letters from all letters
-              const keepLetters = substract(userLetters, putLetters);
-              const updatedLetters = {
-                ...currentGame.letters,
-                [currentUser.id]: keepLetters,
-              };
-              // copy game board to previous board
-              // change game phase to validation
-              // return passedCount to 0
-              await currentGame.update({
-                previousBoard: currentGame.board,
-                board: newBoard,
-                phase: "validation",
-                letters: updatedLetters,
-                // putLetters: putLetters,
-                previousLetters: currentGame.letters[currentUser.id],
-                passedCount: 0,
-                validated: "unknown",
-                lettersChanged: false,
-              });
+                // extract put letters from all letters
+                const keepLetters = substract(userLetters, putLetters);
+                const updatedLetters = {
+                  ...currentGame.letters,
+                  [currentUser.id]: keepLetters,
+                };
+                // copy game board to previous board
+                // change game phase to validation
+                // return passedCount to 0
+                await currentGame.update({
+                  previousBoard: currentGame.board,
+                  board: newBoard,
+                  phase: "validation",
+                  letters: updatedLetters,
+                  // putLetters: putLetters,
+                  previousLetters: currentGame.letters[currentUser.id],
+                  passedCount: 0,
+                  validated: "unknown",
+                  lettersChanged: false,
+                });
+              }
             }
           }
+          // }
         }
       }
       // fetch game from db
@@ -743,7 +475,7 @@ function factory(gameStream, roomStream) {
       });
       const streamAction = {
         type: "GAME_UPDATED",
-        payload: { gameId: gameId, game: updatedGame },
+        payload: { gameId, game: updatedGame },
       };
       const responseAction = {
         type: "NO_DUPLICATIONS",
@@ -823,7 +555,7 @@ function factory(gameStream, roomStream) {
       const action = {
         type: "GAME_UPDATED",
         payload: {
-          gameId: gameId,
+          gameId,
           game: updatedGame,
         },
       };
@@ -889,7 +621,7 @@ function factory(gameStream, roomStream) {
       const action = {
         type: "GAME_UPDATED",
         payload: {
-          gameId: gameId,
+          gameId,
           game: updatedGame,
         },
       };
@@ -961,7 +693,7 @@ function factory(gameStream, roomStream) {
       const action = {
         type: "GAME_UPDATED",
         payload: {
-          gameId: gameId,
+          gameId,
           game: updatedGame,
         },
       };
